Aggiunto redirect a giorno/weekend da table.Fixato problema recupera film da mongoDb
