+++ client/src/components/BoxofficeList.css
@@ -33,3 +33,6 @@ tr {
 td {
   text-align: center;
 }
+thead {
+  text-align: center;
+}
